ADDED -nocleanup flag to all scripts

+++ functions/01_proc-dwi.sh
@@ -22,6 +22,7 @@ BIDS=$1
 id=$2
 out=$3
 PROC=$4
+nocleanup=$5
 here=`pwd`
 
 #------------------------------------------------------------------------------#
@@ -285,7 +286,7 @@ fi
 
 # -----------------------------------------------------------------------------------------------
 # Clean temporal directory
-Do_cmd rm -rf $tmp
+if [ ! $nocleanup = "TRUE" ]; then Do_cmd rm -rf $tmp; fi
 
 # QC notification of completition
 lopuu=$(date +%s)
